+++ CeretMobileTAsk/Domain/Entities/Car.swift
@@ -17,7 +17,7 @@ struct Car {
     let name: String
     let image: String
     let ModelYear: String
-    let startFrom: Int
+    let startFrom: Double
     let identificationAttributeId: Int
     let identificationAttributeValueId: Int
 }
